requirements.txt
streamlit
google-generativeai
pypdf2
streamlit-mic-recorder